tests/wpt/web-platform-tests/css/css-shapes/shape-outside-invalid-circle-003.html

<!doctype html>
<html>
<head>
    <meta charset="UTF-8">
    <meta name="flags" content="dom">
    <meta name="assert" content="shape-outside declaration is invalid if circle shape function has an invalid cx param">
    <title>CSS Shapes Test: Circle - invalid cx value</title>
    <link rel="author" title="Sherlock Zhong" href="mailto:sherlock.z0917@gmail.com">
    <link rel="reviewer" title="Alan Stearns" href="mailto:stearns@adobe.com"> <!-- 11-09-2013 TestTWF Shenzhen -->
    <link rel="help" href="http://www.w3.org/TR/css-shapes-1/#supported-basic-shapes">
    <link rel="help" href="http://www.w3.org/TR/css-shapes-1/#funcdef-ellipse">
    <link rel="help" href="http://www.w3.org/TR/css-shapes-1/#funcdef-circle">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <style>
#shape{
    width: 300px;
    height: 300px;
    shape-outside: circle(50, 50%);
}
    </style>
</head>
<body>

    <div id="log"></div>

    <div id="shape"></div>

    <script>
    var obj = document.getElementById('shape');
    var cx = getComputedStyle(obj).shapeOutside;
    test(function() {assert_equals(cx, "none", "declaration should be invalid")});
    </script>

</body>
</html>
